[FEAT] Proper handling of encrypted data

+++ lib/pfcp/types.h
@@ -1547,9 +1547,9 @@ typedef struct ogs_pfcp_blockchain_data_s {
     uint8_t login[OGS_PFCP_MAX_LOGIN_LEN];
     uint8_t password[OGS_PFCP_MAX_PASSWORD_LEN];
     uint8_t blockchain_node_id[OGS_PFCP_MAX_BLOCKCHAIN_NODE_ID_LEN];
-    uint32_t login_len;
-    uint32_t password_len;
-    uint32_t blockchain_node_id_len;
+    uint8_t login_len;
+    uint16_t password_len;
+    uint8_t blockchain_node_id_len;
 } ogs_pfcp_blockchain_data_t;
 
 int16_t ogs_pfcp_build_volume_measurement(ogs_tlv_octet_t *octet,
